New API

+++ xFunc.Maths/Expressions/Trigonometric/Arccsc.cs
@@ -65,7 +65,7 @@ namespace xFunc.Maths.Expressions.Trigonometric
         /// <seealso cref="ExpressionParameters" />
         protected override double CalculateDergee(ExpressionParameters parameters)
         {
-            return MathExtentions.Acsc(argument.Calculate(parameters)) / Math.PI * 180;
+            return MathExtentions.Acsc((double)argument.Calculate(parameters)) / Math.PI * 180;
         }
 
         /// <summary>
@@ -78,7 +78,7 @@ namespace xFunc.Maths.Expressions.Trigonometric
         /// <seealso cref="ExpressionParameters" />
         protected override double CalculateRadian(ExpressionParameters parameters)
         {
-            return MathExtentions.Acsc(argument.Calculate(parameters));
+            return MathExtentions.Acsc((double)argument.Calculate(parameters));
         }
 
         /// <summary>
@@ -91,7 +91,7 @@ namespace xFunc.Maths.Expressions.Trigonometric
         /// <seealso cref="ExpressionParameters" />
         protected override double CalculateGradian(ExpressionParameters parameters)
         {
-            return MathExtentions.Acsc(argument.Calculate(parameters)) / Math.PI * 200;
+            return MathExtentions.Acsc((double)argument.Calculate(parameters)) / Math.PI * 200;
         }
 
         /// <summary>
