+++ xFunc.Maths/Expressions/Trigonometric/Csc.cs
@@ -68,7 +68,7 @@ namespace xFunc.Maths.Expressions.Trigonometric
         /// <seealso cref="ExpressionParameters" />
         protected override double CalculateDergee(ExpressionParameters parameters)
         {
-            var radian = argument.Calculate(parameters) * Math.PI / 180;
+            var radian = (double)argument.Calculate(parameters) * Math.PI / 180;
 
             return MathExtentions.Csc(radian);
         }
@@ -83,7 +83,7 @@ namespace xFunc.Maths.Expressions.Trigonometric
         /// <seealso cref="ExpressionParameters" />
         protected override double CalculateRadian(ExpressionParameters parameters)
         {
-            return MathExtentions.Csc(argument.Calculate(parameters));
+            return MathExtentions.Csc((double)argument.Calculate(parameters));
         }
 
         /// <summary>
@@ -96,7 +96,7 @@ namespace xFunc.Maths.Expressions.Trigonometric
         /// <seealso cref="ExpressionParameters" />
         protected override double CalculateGradian(ExpressionParameters parameters)
         {
-            var radian = argument.Calculate(parameters) * Math.PI / 200;
+            var radian = (double)argument.Calculate(parameters) * Math.PI / 200;
 
             return MathExtentions.Csc(radian);
         }
